+++ Elements/src/Model.cs
@@ -181,13 +181,23 @@ namespace Elements
         }
 
         /// <summary>
-        /// Get all entities of the specified Type.
+        /// Get all elements of the specified Type.
         /// </summary>
-        /// <typeparam name="T">The Type of element to return.</typeparam>
+        /// <typeparam name="T">The type of element to return.</typeparam>
         /// <returns>A collection of elements of the specified type.</returns>
         public IEnumerable<T> AllElementsOfType<T>()
         {
             return this.Elements.Values.OfType<T>();
+        }
+
+        /// <summary>
+        /// Get all elements derived from the specified type.
+        /// </summary>
+        /// <typeparam name="T">The type of the element from which returned elements derive.</typeparam>
+        /// <returns>A collection of elements derived from the specified type.</returns>
+        public IEnumerable<T> AllElementsDerivedFromType<T>()
+        {
+            return this.Elements.Values.Where(e => typeof(T).IsAssignableFrom(e.GetType())).Cast<T>();
         }
 
         /// <summary>