+++ Elements/test/ModelTests.cs
@@ -353,6 +353,17 @@ namespace Elements.Tests
             Assert.False(Model.IsValidForRecursiveAddition(typeof(object)));
         }
 
+        [Fact]
+        public void AllElementsDerivedFromType()
+        {
+            var column = new Column(new Vector3(5, 5, 5), 2.0, Polygon.Rectangle(1, 1));
+            var beam = new Beam(new Line(Vector3.Origin, new Vector3(5, 5, 5)), Polygon.Rectangle(1, 1));
+            var brace = new Brace(new Line(Vector3.Origin, new Vector3(5, 5, 5)), Polygon.Rectangle(1, 1));
+            var model = new Model();
+            model.AddElements(column, beam, brace);
+            Assert.Equal(3, model.AllElementsDerivedFromType<StructuralFraming>().Count());
+        }
+
         private Model QuadPanelModel()
         {
             var model = new Model();